src/mcp-search/requirements.txt
fastapi
uvicorn[standard]
azure-identity
azure-search-documents
